Move Benchmarks into separate file
